requirements.txt
mcp[cli]
python-dotenv
openai
ipykernel
httpx
nest-asyncio